+++ pkg/virt-api/webhooks/validating-webhook/admitters/BUILD.bazel
@@ -7,7 +7,6 @@ go_library(
         "migration-create-admitter.go",
         "migration-update-admitter.go",
         "migrationpolicy-admitter.go",
-        "network.go",
         "pod-eviction-admitter.go",
         "preference-admitter.go",
         "status-admitter.go",
@@ -31,8 +30,8 @@ go_library(
         "//pkg/hooks:go_default_library",
         "//pkg/instancetype:go_default_library",
         "//pkg/monitoring/metrics/virt-api:go_default_library",
+        "//pkg/network/admitter:go_default_library",
         "//pkg/network/link:go_default_library",
-        "//pkg/network/vmispec:go_default_library",
         "//pkg/storage/backend-storage:go_default_library",
         "//pkg/storage/reservation:go_default_library",
         "//pkg/storage/snapshot:go_default_library",
@@ -88,7 +87,6 @@ go_test(
         "migration-create-admitter_test.go",
         "migration-update-admitter_test.go",
         "migrationpolicy-admitter_test.go",
-        "network_test.go",
         "pod-eviction-admitter_test.go",
         "preference-admitter_test.go",
         "vmclone-admitter_test.go",
